mvc cleanup and updates to mvc/api tests
