--- src/test/java/emissary/server/mvc/ShutdownActionTest.java
@@ -1,14 +1,0 @@
-package emissary.server.mvc;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-class ShutdownActionTest extends EndpointTestBase {
-
-    @Disabled("Not yet implemented")
-    @Test
-    void test() {
-        // fail("Not yet implemented");
-    }
-
-}
